CACHED_DB/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
duckdb==1.4.4
mysql-connector-python==9.5.0
pandas==2.3.3
numpy==2.2.6
openpyxl==3.1.5
XlsxWriter==3.2.3
python-multipart==0.0.20
